Move cache adapters into their own module
